4.1

+++ draw_ray.c
@@ -13,7 +13,7 @@
 #include "header/my_type.h"
 #include "mlx/mlx.h"
 
-void	draw_ray(int pov, t_player player, char (*map)[10], t_data *img)
+void	draw_ray(double pov, t_player player, char (*map)[10], t_data *img)
 {
 	t_collis	collis;
 	float		phi;
@@ -35,10 +35,14 @@ void	draw_ray(int pov, t_player player, char (*map)[10], t_data *img)
 		collis = find_collision(new_phi * M_PI / 180, player, map);
 		print_line(player.x, player.y, collis.x, collis.y, img);
 		phi -= d_phi;
+		// if (collis.x > 640 || collis.x < 0 || collis.y > 640 || collis.y < 0)
+		// {
+		// 	printf("x = %f  y = %f   pov = %f\n", collis.x, collis.y, pov);
+		// }
 	}
 }
 
-// void	draw_ray(int pov, t_player player, char (*map)[10], t_data *img)
+// void	draw_ray(double pov, t_player player, char (*map)[10], t_data *img)
 // {
 // 	t_collis	collis;
 // 	float		phi;
@@ -48,6 +52,10 @@ void	draw_ray(int pov, t_player player, char (*map)[10], t_data *img)
 
 // 	d_phi = 66.0 / 1920.0;
 // 	collis = find_collision(pov * M_PI / 180, player, map);
-// 	printf("pov = %d\n", pov);
-// 	print_line(player.x, player.y, collis.x, collis.y, img);
+// 	if (collis.x > 640 || collis.x < 0 || collis.y > 640 || collis.y < 0)
+// 	{
+// 		printf("x = %f  y = %f   pov = %f\n", collis.x, collis.y, pov);
+// 	}
+// 	// printf("pov = %f\n", pov);
+// 	// print_line(player.x, player.y, collis.x, collis.y, img);
 // }
